huff0: add arm64 assembly for Decompress4X/1X via avo lowering
Generate arm64 asm from the existing avo source with the honeycombio/avo
arm64 lowering printer (-arch amd64,arm64), the same mechanism zstd's
seqdec/fse asm already uses. The amd64 output is unchanged.

The Go-side wrappers move to decompress_asm.go, shared between amd64 and
arm64; per-arch files keep only the asm stubs and dispatch (amd64 retains
the BMI2 selection for Decompress1X; arm64 lowers the generic twin).

The avo pin gains exact x86 partial-register semantics (MOVB/MOVW/ADDB/
ADCB/SETcc lower to bit-field inserts preserving untouched bits) — the
huff0 loops pack output bytes through AH and accumulate byte counters,
which the previous lowering silently corrupted — plus constant-folding of
BMI2 control registers (BEXTR with an adjacent constant control becomes a
single UBFX). zstd/seqdec_arm64.s is regenerated with the same printer;
its amd64 output is byte-identical to before.

On a 16-core Cortex-A72 (ARMv8.0), huff0 Decompress1X improves 25-34%
(twain 114->153 MB/s, digits 128->161 MB/s); Decompress4X is content-
dependent (+12% twain, -4% low-entropy). zstd DecodeAll improves 1-7%
end-to-end (urls.10K +3.8%, html +2.2%, comp-data.bin +6.6%), with
checksum (18% of CPU, at scalar xxhash roofline) and sequence decoding
dominating the remainder.

Validated: full huff0 + zstd test suites on arm64 (Cortex-A72) and amd64;
avo differential execution tests for every new/changed lowering on both
architectures.
