+++ huff0/_generate/go.mod
@@ -10,9 +10,11 @@ require (
 )
 
 require (
-	golang.org/x/mod v0.21.0 // indirect
-	golang.org/x/sync v0.8.0 // indirect
-	golang.org/x/tools v0.25.1 // indirect
+	golang.org/x/mod v0.27.0 // indirect
+	golang.org/x/sync v0.16.0 // indirect
+	golang.org/x/tools v0.36.0 // indirect
 )
 
 replace github.com/klauspost/compress => ../..
+
+replace github.com/mmcloughlin/avo => github.com/honeycombio/avo v0.6.1-0.20260718041253-a4dbeacc0c55
