+++ zstd/_generate/go.mod
@@ -17,4 +17,4 @@ require (
 
 replace github.com/klauspost/compress => ../..
 
-replace github.com/mmcloughlin/avo => github.com/honeycombio/avo v0.6.1-0.20260709170202-520357590ac6
+replace github.com/mmcloughlin/avo => github.com/honeycombio/avo v0.6.1-0.20260718041253-a4dbeacc0c55
